libavformat/mpegts.c: add: parse EIT descriptors
This patch add to parse descriptors on EIT packets.
The patch is intended to set information to current program and/or
A/V stream.
On Japanese ISDB, some important / useful information is provided
via EIT only.
ref: ARIB STD B10 Table 6-1, Section 6.1, Part 2.

This patch only parse short event descriptor (0x4d) and set title
information to the program. It may not be useful, but a good example
of EIT descriptor common for DVB and ISDB.

Note that it only parse EIT for actual and present TS stream as bllow:
* Parse EIT table_id 0x4E (actual TS stream) only.
  ref: DVB Blue Book A038 (EN 300 468) Table 2, Section 5.1.3.
* Parse section number 0x00 only.
  Section number 0x00 is present event.
  Section number 0x01 is following (i.e., not for present stream).
  Section number 0x02 or later may contain event for present stream,
  but it is hard to distinguish and rarely sent.
  ref: DVB-SI Guidelines (TS 101 211) Section 4.1.4.1.
* Find a program associated to the EIT in already initialized
  AVProgram and Program structures.
  If no program found, abort to parse the EIT.

Since EIT packets may be sent several times for the same program,
add 'cur_eit_ver' in Program structure and ignore EITs with the
same version as previously parsed one.

There is a warning: "variable 'language' set but not used" at
compilation. It should be resolved by future patches.